+++ cli/tests/test_operations.rs
@@ -507,6 +507,608 @@ fn test_op_abandon_without_updating_working_copy() {
     "###);
 }
 
+#[test]
+fn test_op_diff() {
+    let test_env = TestEnvironment::default();
+    let git_repo_path = test_env.env_root().join("git-repo");
+    let git_repo = init_bare_git_repo(&git_repo_path);
+    test_env.jj_cmd_ok(test_env.env_root(), &["git", "clone", "git-repo", "repo"]);
+    let repo_path = test_env.env_root().join("repo");
+
+    // Overview of op log.
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "log"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    @  984d5ceb039f test-username@host.example.com 2001-02-03 04:05:07.000 +07:00 - 2001-02-03 04:05:07.000 +07:00
+    │  check out git remote's default branch
+    │  args: jj git clone git-repo repo
+    ○  817baaeefcbb test-username@host.example.com 2001-02-03 04:05:07.000 +07:00 - 2001-02-03 04:05:07.000 +07:00
+    │  fetch from git remote into empty repo
+    │  args: jj git clone git-repo repo
+    ○  b51416386f26 test-username@host.example.com 2001-02-03 04:05:07.000 +07:00 - 2001-02-03 04:05:07.000 +07:00
+    │  add workspace 'default'
+    ○  9a7d829846af test-username@host.example.com 2001-02-03 04:05:07.000 +07:00 - 2001-02-03 04:05:07.000 +07:00
+    │  initialize repo
+    ○  000000000000 root()
+    "###);
+
+    // Diff between the same operation should be empty.
+    let stdout = test_env.jj_cmd_success(
+        &repo_path,
+        &["op", "diff", "--from", "0000000", "--to", "0000000"],
+    );
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 000000000000: root()
+      To operation 000000000000: root()
+
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff", "--from", "@", "--to", "@"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 984d5ceb039f: check out git remote's default branch
+      To operation 984d5ceb039f: check out git remote's default branch
+    "###);
+
+    // Diff from parent operation to latest operation.
+    // `jj op diff --op @` should behave identically to `jj op diff --from
+    // @- --to @` (if `@` is not a merge commit).
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff", "--from", "@-", "--to", "@"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 817baaeefcbb: fetch from git remote into empty repo
+      To operation 984d5ceb039f: check out git remote's default branch
+
+    Changed commits:
+    ○  Change sqpuoqvxutmz
+       + sqpuoqvx 9708515f (empty) (no description set)
+    ○  Change qpvuntsmwlqt
+       - qpvuntsm hidden 230dd059 (empty) (no description set)
+
+    Changed local branches:
+    branch-1:
+    + ulyvmwyz 1d843d1f branch-1 | Commit 1
+    - (absent)
+
+    Changed remote branches:
+    branch-1@origin:
+    + tracked ulyvmwyz 1d843d1f branch-1 | Commit 1
+    - untracked ulyvmwyz 1d843d1f branch-1 | Commit 1
+    "###);
+    let stdout_without_from_to = test_env.jj_cmd_success(&repo_path, &["op", "diff"]);
+    assert_eq!(stdout, stdout_without_from_to);
+
+    // Diff from root operation to latest operation
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff", "--from", "0000000"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 000000000000: root()
+      To operation 984d5ceb039f: check out git remote's default branch
+
+    Changed commits:
+    ○  Change sqpuoqvxutmz
+    │  + sqpuoqvx 9708515f (empty) (no description set)
+    ○  Change ulyvmwyzwuwt
+    │  + ulyvmwyz 1d843d1f branch-1 | Commit 1
+    │ ○  Change tqyxmsztkvot
+    ├─╯  + tqyxmszt 3e785984 branch-3@origin | Commit 3
+    │ ○  Change yuvsmzqkmpws
+    ├─╯  + yuvsmzqk 3d9189bc branch-2@origin | Commit 2
+    ○  Change zzzzzzzzzzzz
+       + zzzzzzzz 00000000 (empty) (no description set)
+
+    Changed local branches:
+    branch-1:
+    + ulyvmwyz 1d843d1f branch-1 | Commit 1
+    - (absent)
+
+    Changed remote branches:
+    branch-1@origin:
+    + tracked ulyvmwyz 1d843d1f branch-1 | Commit 1
+    - untracked (absent)
+    branch-2@origin:
+    + untracked yuvsmzqk 3d9189bc branch-2@origin | Commit 2
+    - untracked (absent)
+    branch-3@origin:
+    + untracked tqyxmszt 3e785984 branch-3@origin | Commit 3
+    - untracked (absent)
+    "###);
+
+    // Diff from latest operation to root operation
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff", "--to", "0000000"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 984d5ceb039f: check out git remote's default branch
+      To operation 000000000000: root()
+
+    Changed commits:
+    ○  Change sqpuoqvxutmz
+    │  - sqpuoqvx hidden 9708515f (empty) (no description set)
+    ○  Change ulyvmwyzwuwt
+    │  - ulyvmwyz hidden 1d843d1f Commit 1
+    │ ○  Change tqyxmsztkvot
+    ├─╯  - tqyxmszt hidden 3e785984 Commit 3
+    │ ○  Change yuvsmzqkmpws
+    ├─╯  - yuvsmzqk hidden 3d9189bc Commit 2
+    ○  Change zzzzzzzzzzzz
+       - zzzzzzzz hidden 00000000 (empty) (no description set)
+
+    Changed local branches:
+    branch-1:
+    + (absent)
+    - ulyvmwyz hidden 1d843d1f Commit 1
+
+    Changed remote branches:
+    branch-1@origin:
+    + untracked (absent)
+    - tracked ulyvmwyz hidden 1d843d1f Commit 1
+    branch-2@origin:
+    + untracked (absent)
+    - untracked yuvsmzqk hidden 3d9189bc Commit 2
+    branch-3@origin:
+    + untracked (absent)
+    - untracked tqyxmszt hidden 3e785984 Commit 3
+    "###);
+
+    // Create a conflicted branch using a concurrent operation.
+    test_env.jj_cmd_ok(
+        &repo_path,
+        &[
+            "branch",
+            "set",
+            "branch-1",
+            "-r",
+            "branch-2@origin",
+            "--at-op",
+            "@-",
+        ],
+    );
+    let (_, stderr) = test_env.jj_cmd_ok(&repo_path, &["log"]);
+    insta::assert_snapshot!(&stderr, @r###"
+    Concurrent modification detected, resolving automatically.
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "log"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    @    6eeb006eccd0 test-username@host.example.com 2001-02-03 04:05:16.000 +07:00 - 2001-02-03 04:05:16.000 +07:00
+    ├─╮  resolve concurrent operations
+    │ │  args: jj log
+    ○ │  984d5ceb039f test-username@host.example.com 2001-02-03 04:05:07.000 +07:00 - 2001-02-03 04:05:07.000 +07:00
+    │ │  check out git remote's default branch
+    │ │  args: jj git clone git-repo repo
+    │ ○  5ed581429582 test-username@host.example.com 2001-02-03 04:05:15.000 +07:00 - 2001-02-03 04:05:15.000 +07:00
+    ├─╯  point branch branch-1 to commit 3d9189bc56a1972729350456eb95ec5bf90be2a8
+    │    args: jj branch set branch-1 -r branch-2@origin --at-op @-
+    ○  817baaeefcbb test-username@host.example.com 2001-02-03 04:05:07.000 +07:00 - 2001-02-03 04:05:07.000 +07:00
+    │  fetch from git remote into empty repo
+    │  args: jj git clone git-repo repo
+    ○  b51416386f26 test-username@host.example.com 2001-02-03 04:05:07.000 +07:00 - 2001-02-03 04:05:07.000 +07:00
+    │  add workspace 'default'
+    ○  9a7d829846af test-username@host.example.com 2001-02-03 04:05:07.000 +07:00 - 2001-02-03 04:05:07.000 +07:00
+    │  initialize repo
+    ○  000000000000 root()
+    "###);
+    let op_log_lines = stdout.lines().collect_vec();
+    let op_id = op_log_lines[0].split(' ').nth(4).unwrap();
+    let first_parent_id = op_log_lines[3].split(' ').nth(3).unwrap();
+    let second_parent_id = op_log_lines[6].split(' ').nth(3).unwrap();
+
+    // Diff between the first parent of the merge operation and the merge operation.
+    let stdout = test_env.jj_cmd_success(
+        &repo_path,
+        &["op", "diff", "--from", first_parent_id, "--to", op_id],
+    );
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 984d5ceb039f: check out git remote's default branch
+      To operation 6eeb006eccd0: resolve concurrent operations
+
+    Changed local branches:
+    branch-1:
+    + (added) ulyvmwyz 1d843d1f branch-1?? branch-1@origin | Commit 1
+    + (added) yuvsmzqk 3d9189bc branch-1?? branch-2@origin | Commit 2
+    - ulyvmwyz 1d843d1f branch-1?? branch-1@origin | Commit 1
+    "###);
+
+    // Diff between the second parent of the merge operation and the merge
+    // operation.
+    let stdout = test_env.jj_cmd_success(
+        &repo_path,
+        &["op", "diff", "--from", second_parent_id, "--to", op_id],
+    );
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 5ed581429582: point branch branch-1 to commit 3d9189bc56a1972729350456eb95ec5bf90be2a8
+      To operation 6eeb006eccd0: resolve concurrent operations
+
+    Changed commits:
+    ○  Change sqpuoqvxutmz
+       + sqpuoqvx 9708515f (empty) (no description set)
+    ○  Change qpvuntsmwlqt
+       - qpvuntsm hidden 230dd059 (empty) (no description set)
+
+    Changed local branches:
+    branch-1:
+    + (added) ulyvmwyz 1d843d1f branch-1?? branch-1@origin | Commit 1
+    + (added) yuvsmzqk 3d9189bc branch-1?? branch-2@origin | Commit 2
+    - yuvsmzqk 3d9189bc branch-1?? branch-2@origin | Commit 2
+
+    Changed remote branches:
+    branch-1@origin:
+    + tracked ulyvmwyz 1d843d1f branch-1?? branch-1@origin | Commit 1
+    - untracked ulyvmwyz 1d843d1f branch-1?? branch-1@origin | Commit 1
+    "###);
+
+    // Test fetching from git remote.
+    modify_git_repo(git_repo);
+    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["git", "fetch"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    "###);
+    insta::assert_snapshot!(&stderr, @r###"
+    branch: branch-1@origin [updated] tracked
+    branch: branch-2@origin [updated] untracked
+    branch: branch-3@origin [deleted] untracked
+    Abandoned 1 commits that are no longer reachable.
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 6eeb006eccd0: resolve concurrent operations
+      To operation 9c57642e4a18: fetch from git remote(s) origin
+
+    Changed commits:
+    ○  Change qzxslznxxpoz
+       + qzxslznx d487febd branch-2@origin | Commit 5
+    ○  Change slvtnnzxztqy
+       + slvtnnzx 4f856199 branch-1?? branch-1@origin | Commit 4
+    ○  Change tqyxmsztkvot
+       - tqyxmszt hidden 3e785984 Commit 3
+
+    Changed local branches:
+    branch-1:
+    + (added) slvtnnzx 4f856199 branch-1?? branch-1@origin | Commit 4
+    + (added) yuvsmzqk 3d9189bc branch-1?? | Commit 2
+    - (added) ulyvmwyz 1d843d1f Commit 1
+    - (added) yuvsmzqk 3d9189bc branch-1?? | Commit 2
+
+    Changed remote branches:
+    branch-1@origin:
+    + tracked slvtnnzx 4f856199 branch-1?? branch-1@origin | Commit 4
+    - tracked ulyvmwyz 1d843d1f Commit 1
+    branch-2@origin:
+    + untracked qzxslznx d487febd branch-2@origin | Commit 5
+    - untracked yuvsmzqk 3d9189bc branch-1?? | Commit 2
+    branch-3@origin:
+    + untracked (absent)
+    - untracked tqyxmszt hidden 3e785984 Commit 3
+    "###);
+
+    // Test creation of branch.
+    let (stdout, stderr) = test_env.jj_cmd_ok(
+        &repo_path,
+        &["branch", "create", "branch-2", "-r", "branch-2@origin"],
+    );
+    insta::assert_snapshot!(&stdout, @r###"
+    "###);
+    insta::assert_snapshot!(&stderr, @r###"
+    Created 1 branches pointing to qzxslznx d487febd branch-2 branch-2@origin | Commit 5
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 9c57642e4a18: fetch from git remote(s) origin
+      To operation 8b280b4a5ea2: create branch branch-2 pointing to commit d487febd08e690ee775a4e0387e30d544307e409
+
+    Changed local branches:
+    branch-2:
+    + qzxslznx d487febd branch-2 branch-2@origin | Commit 5
+    - (absent)
+    "###);
+
+    // Test tracking of branch.
+    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["branch", "track", "branch-2@origin"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    "###);
+    insta::assert_snapshot!(&stderr, @r###"
+    Started tracking 1 remote branches.
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 8b280b4a5ea2: create branch branch-2 pointing to commit d487febd08e690ee775a4e0387e30d544307e409
+      To operation be38bc6501bc: track remote branch branch-2@origin
+
+    Changed remote branches:
+    branch-2@origin:
+    + tracked qzxslznx d487febd branch-2 | Commit 5
+    - untracked qzxslznx d487febd branch-2 | Commit 5
+    "###);
+
+    // Test creation of new commit.
+    let (stdout, stderr) =
+        test_env.jj_cmd_ok(&repo_path, &["new", "branch-1@origin", "-m", "new commit"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    "###);
+    insta::assert_snapshot!(&stderr, @r###"
+    Working copy now at: nmzmmopx bed2698f (empty) new commit
+    Parent commit      : slvtnnzx 4f856199 branch-1?? branch-1@origin | Commit 4
+    Added 1 files, modified 0 files, removed 1 files
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation be38bc6501bc: track remote branch branch-2@origin
+      To operation 8c9091fb718a: new empty commit
+
+    Changed commits:
+    ○  Change nmzmmopxokps
+       + nmzmmopx bed2698f (empty) new commit
+    ○  Change sqpuoqvxutmz
+       - sqpuoqvx hidden 9708515f (empty) (no description set)
+    "###);
+
+    // Test updating of local branch.
+    let (stdout, stderr) =
+        test_env.jj_cmd_ok(&repo_path, &["branch", "set", "branch-1", "-r", "@"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    "###);
+    insta::assert_snapshot!(&stderr, @r###"
+    Moved 1 branches to nmzmmopx bed2698f branch-1* | (empty) new commit
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 8c9091fb718a: new empty commit
+      To operation 6ff61c177324: point branch branch-1 to commit bed2698f6baf06f7eea56c616bc3fe36d9065651
+
+    Changed local branches:
+    branch-1:
+    + nmzmmopx bed2698f branch-1* | (empty) new commit
+    - (added) slvtnnzx 4f856199 branch-1@origin | Commit 4
+    - (added) yuvsmzqk 3d9189bc Commit 2
+    "###);
+
+    // Test deletion of local branch.
+    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["branch", "delete", "branch-2"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    "###);
+    insta::assert_snapshot!(&stderr, @r###"
+    Deleted 1 branches.
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 6ff61c177324: point branch branch-1 to commit bed2698f6baf06f7eea56c616bc3fe36d9065651
+      To operation ecae5e879b40: delete branch branch-2
+
+    Changed local branches:
+    branch-2:
+    + (absent)
+    - qzxslznx d487febd branch-2@origin | Commit 5
+    "###);
+
+    // Test pushing to Git remote.
+    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["git", "push", "--tracked"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    "###);
+    insta::assert_snapshot!(&stderr, @r###"
+    Branch changes to push to origin:
+      Move forward branch branch-1 from 4f856199edbf to bed2698f6baf
+      Delete branch branch-2 from d487febd08e6
+    Warning: The working-copy commit in workspace 'default' became immutable, so a new commit has been created on top of it.
+    Working copy now at: uuuvxpvw 2c8e84a8 (empty) (no description set)
+    Parent commit      : nmzmmopx bed2698f branch-1 | (empty) new commit
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation ecae5e879b40: delete branch branch-2
+      To operation 96f11847b661: push all tracked branches to git remote origin
+
+    Changed commits:
+    ○  Change uuuvxpvwspwr
+       + uuuvxpvw 2c8e84a8 (empty) (no description set)
+
+    Changed remote branches:
+    branch-1@origin:
+    + tracked nmzmmopx bed2698f branch-1 | (empty) new commit
+    - tracked slvtnnzx 4f856199 Commit 4
+    branch-2@origin:
+    + untracked (absent)
+    - tracked qzxslznx d487febd Commit 5
+    "###);
+}
+
+#[test]
+fn test_op_diff_patch() {
+    let test_env = TestEnvironment::default();
+    test_env.jj_cmd_ok(test_env.env_root(), &["git", "init", "repo"]);
+    let repo_path = test_env.env_root().join("repo");
+
+    // Update working copy with a single file and create new commit.
+    std::fs::write(repo_path.join("file"), "a\n").unwrap();
+    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["new"]);
+    insta::assert_snapshot!(&stdout, @"");
+    insta::assert_snapshot!(&stderr, @r###"
+    Working copy now at: rlvkpnrz 56950632 (empty) (no description set)
+    Parent commit      : qpvuntsm 6b1027d2 (no description set)
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff", "--op", "@-", "-p", "--git"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation b51416386f26: add workspace 'default'
+      To operation 6188e9d1f7da: snapshot working copy
+
+    Changed commits:
+    ○  Change qpvuntsmwlqt
+       + qpvuntsm 6b1027d2 (no description set)
+       - qpvuntsm hidden 230dd059 (empty) (no description set)
+       diff --git a/file b/file
+       new file mode 100644
+       index 0000000000..7898192261
+       --- /dev/null
+       +++ b/file
+       @@ -1,0 +1,1 @@
+       +a
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff", "--op", "@", "-p", "--git"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 6188e9d1f7da: snapshot working copy
+      To operation 8f6a879bef11: new empty commit
+
+    Changed commits:
+    ○  Change rlvkpnrzqnoo
+       + rlvkpnrz 56950632 (empty) (no description set)
+    "###);
+
+    // Squash the working copy commit.
+    std::fs::write(repo_path.join("file"), "b\n").unwrap();
+    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["squash"]);
+    insta::assert_snapshot!(&stdout, @"");
+    insta::assert_snapshot!(&stderr, @r###"
+    Working copy now at: mzvwutvl 9f4fb57f (empty) (no description set)
+    Parent commit      : qpvuntsm 2ac85fd1 (no description set)
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff", "-p", "--git"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation 874d3a8b4c77: snapshot working copy
+      To operation c53f5f1afbc6: squash commits into 6b1027d2770cd0a39c468e525e52bf8c47e1464a
+
+    Changed commits:
+    ○  Change mzvwutvlkqwt
+    │  + mzvwutvl 9f4fb57f (empty) (no description set)
+    │ ○  Change rlvkpnrzqnoo
+    ├─╯  - rlvkpnrz hidden 1d7f8f94 (no description set)
+    │    diff --git a/file b/file
+    │    index 7898192261..6178079822 100644
+    │    --- a/file
+    │    +++ b/file
+    │    @@ -1,1 +1,1 @@
+    │    -a
+    │    +b
+    ○  Change qpvuntsmwlqt
+       + qpvuntsm 2ac85fd1 (no description set)
+       - qpvuntsm hidden 6b1027d2 (no description set)
+       diff --git a/file b/file
+       index 7898192261..6178079822 100644
+       --- a/file
+       +++ b/file
+       @@ -1,1 +1,1 @@
+       -a
+       +b
+    "###);
+
+    // Abandon the working copy commit.
+    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["abandon"]);
+    insta::assert_snapshot!(&stdout, @"");
+    insta::assert_snapshot!(&stderr, @r###"
+    Abandoned commit mzvwutvl 9f4fb57f (empty) (no description set)
+    Working copy now at: yqosqzyt 33f321c4 (empty) (no description set)
+    Parent commit      : qpvuntsm 2ac85fd1 (no description set)
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "diff", "-p", "--git"]);
+    insta::assert_snapshot!(&stdout, @r###"
+    From operation c53f5f1afbc6: squash commits into 6b1027d2770cd0a39c468e525e52bf8c47e1464a
+      To operation e13dc1c7a3b3: abandon commit 9f4fb57fba25a7b47ce5980a5d9a4766778331e8
+
+    Changed commits:
+    ○  Change yqosqzytrlsw
+       + yqosqzyt 33f321c4 (empty) (no description set)
+    ○  Change mzvwutvlkqwt
+       - mzvwutvl hidden 9f4fb57f (empty) (no description set)
+    "###);
+}
+
+fn init_bare_git_repo(git_repo_path: &Path) -> git2::Repository {
+    let git_repo = git2::Repository::init_bare(git_repo_path).unwrap();
+    let git_blob_oid = git_repo.blob(b"some content").unwrap();
+    let mut git_tree_builder = git_repo.treebuilder(None).unwrap();
+    git_tree_builder
+        .insert("some-file", git_blob_oid, 0o100644)
+        .unwrap();
+    let git_tree_id = git_tree_builder.write().unwrap();
+    drop(git_tree_builder);
+    let git_tree = git_repo.find_tree(git_tree_id).unwrap();
+    let git_signature = git2::Signature::new(
+        "Git User",
+        "git.user@example.com",
+        &git2::Time::new(123, 60),
+    )
+    .unwrap();
+    git_repo
+        .commit(
+            Some("refs/heads/branch-1"),
+            &git_signature,
+            &git_signature,
+            "Commit 1",
+            &git_tree,
+            &[],
+        )
+        .unwrap();
+    git_repo
+        .commit(
+            Some("refs/heads/branch-2"),
+            &git_signature,
+            &git_signature,
+            "Commit 2",
+            &git_tree,
+            &[],
+        )
+        .unwrap();
+    git_repo
+        .commit(
+            Some("refs/heads/branch-3"),
+            &git_signature,
+            &git_signature,
+            "Commit 3",
+            &git_tree,
+            &[],
+        )
+        .unwrap();
+    drop(git_tree);
+    git_repo.set_head("refs/heads/branch-1").unwrap();
+    git_repo
+}
+
+fn modify_git_repo(git_repo: git2::Repository) -> git2::Repository {
+    let git_blob_oid = git_repo.blob(b"more content").unwrap();
+    let mut git_tree_builder = git_repo.treebuilder(None).unwrap();
+    git_tree_builder
+        .insert("next-file", git_blob_oid, 0o100644)
+        .unwrap();
+    let git_tree_id = git_tree_builder.write().unwrap();
+    drop(git_tree_builder);
+    let git_tree = git_repo.find_tree(git_tree_id).unwrap();
+    let git_signature = git2::Signature::new(
+        "Git User",
+        "git.user@example.com",
+        &git2::Time::new(123, 60),
+    )
+    .unwrap();
+    let branch1_commit = git_repo
+        .find_reference("refs/heads/branch-1")
+        .unwrap()
+        .peel_to_commit()
+        .unwrap();
+    let branch2_commit = git_repo
+        .find_reference("refs/heads/branch-2")
+        .unwrap()
+        .peel_to_commit()
+        .unwrap();
+    git_repo
+        .commit(
+            Some("refs/heads/branch-1"),
+            &git_signature,
+            &git_signature,
+            "Commit 4",
+            &git_tree,
+            &[&branch1_commit],
+        )
+        .unwrap();
+    git_repo
+        .commit(
+            Some("refs/heads/branch-2"),
+            &git_signature,
+            &git_signature,
+            "Commit 5",
+            &git_tree,
+            &[&branch2_commit],
+        )
+        .unwrap();
+    git_repo
+        .find_reference("refs/heads/branch-3")
+        .unwrap()
+        .delete()
+        .unwrap();
+    drop(git_tree);
+    drop(branch1_commit);
+    drop(branch2_commit);
+    git_repo
+}
+
 fn get_log_output(test_env: &TestEnvironment, repo_path: &Path, op_id: &str) -> String {
     test_env.jj_cmd_success(
         repo_path,